chore: refactor dashboard
