+++ src/App.tsx
@@ -4,18 +4,16 @@ import { WalletModalProvider } from "./providers/WalletModal";
 import ApolloProvider from "./providers/Apollo";
 import { ConfigProvider } from "./providers/Config";
 
-const App = () => {
-  return (
-    <ConfigProvider>
-      <WagmiProvider>
-        <WalletModalProvider>
-          <ApolloProvider>
-            <Navigation />
-          </ApolloProvider>
-        </WalletModalProvider>
-      </WagmiProvider>
-    </ConfigProvider>
-  );
-};
+const App = () => (
+  <ConfigProvider>
+    <WagmiProvider>
+      <WalletModalProvider>
+        <ApolloProvider>
+          <Navigation />
+        </ApolloProvider>
+      </WalletModalProvider>
+    </WagmiProvider>
+  </ConfigProvider>
+);
 
 export default App;
