+++ src/components/Buttons/index.ts
@@ -1,2 +1,3 @@
 export * from "./BaseButton";
 export * from "./ConnectWalletButton";
+export * from "./ClaimTokensButton";
