+++ src/components/Dashboard/DashboardTable.tsx
@@ -1,95 +1,16 @@
+import React from "react";
+import { Meet } from "../../types/meets";
 import { Link } from "react-router-dom";
 import moment from "moment";
 import Skeleton from "react-loading-skeleton";
-import { Toggle, PageLayout } from "../../components";
-import MyStats from "./components/MyStats";
-import OverallStats from "./components/OverallStats";
-import { Meet } from "../../types/meets";
 
 type Props = {
-  asLocaltime: (raceTime: number) => string;
-  meets: Meet[];
-  signature: string | undefined;
-  owner: string | undefined;
-  myPlayEnabled: boolean;
-  onMyPlayToggle: () => void;
-};
-type TableProps = {
-  asLocaltime: (raceTime: number) => string;
   meets: Meet[];
 };
-const DashboardView: React.FC<Props> = (props: Props) => {
-  const {
-    asLocaltime,
-    meets,
-    owner,
-    signature,
-    myPlayEnabled,
-    onMyPlayToggle
-  } = props;
 
-  return (
-    <PageLayout>
-      <div className="grid gap-6">
-        <div>
-          <div className="container-fluid px-4 py-5 bg-emerald-700 shadow rounded-lg overflow-hidden sm:p-6">
-            <div className="flex flex-wrap justify-between">
-              <img
-                loading="lazy"
-                alt="Horse-Link"
-                src="/images/horse-link.png"
-                className="mt-2 mb-8"
-              />
-              <img
-                loading="lazy"
-                alt="Horse"
-                src="/images/horse.png"
-                className="h-20"
-              />
-            </div>
-            <h2 className="text-lg mb-3 font-medium text-gray-900">
-              Horse Link is an Ethereum AMM protocol that allows participants to
-              wager on sports markets using ERC20 tokens.
-            </h2>
-            <p className="text-xs my-2">
-              Horse Link&apos;s smart contract guaranteed bets are always placed
-              within the slippage band of the constant product function. Like
-              other AMM protocols based on curve functions, bets based within
-              the range of slippage based on the payout will be placed.
-            </p>
-          </div>
-          {myPlayEnabled ? <MyStats /> : <OverallStats />}
-        </div>
-        <div className="flex gap-3 self-end justify-self-end">
-          <Toggle enabled={myPlayEnabled} onChange={onMyPlayToggle} />
-          <div>My Stats</div>
-        </div>
-        <div className="-mt-12">
-          <Table asLocaltime={asLocaltime} meets={meets} />
-        </div>
-        <div className="flex justify-center px-4 py-5 bg-white shadow rounded-lg sm:p-6">
-          <div className="w-4/5 max-w-2xl">
-            <div className="flex flex-col items-center">
-              <h2 className="text-lg">Signature :</h2>
-              <h2 className="break-all">
-                {signature || <Skeleton width={"25em"} count={2} />}
-              </h2>
-            </div>
-            <div className="mt-3 flex flex-col items-center">
-              <h2 className="text-lg">Owner Address :</h2>
-              <h2 className="break-all">
-                {owner || <Skeleton width={"25em"} />}
-              </h2>
-            </div>
-          </div>
-        </div>
-      </div>
-    </PageLayout>
-  );
-};
-const Table: React.FC<TableProps> = (props: TableProps) => {
-  const { meets } = props;
+export const DashboardTable: React.FC<Props> = ({ meets }) => {
   const maxLength = Math.max(...meets.map(meet => meet.races.length));
+
   return (
     <div className="grid grid-cols-2">
       <div className="col-span-2">
@@ -177,4 +98,3 @@ const Table: React.FC<TableProps> = (props: TableProps) => {
     </div>
   );
 };
-export default DashboardView;
