+++ src/components/Dashboard/index.ts
@@ -1,0 +1,3 @@
+export * from "./DashboardOverallStats";
+export * from "./DashboardUserStats";
+export * from "./DashboardTable";
