--- src/pages/NotFound/NotFound_View.tsx
@@ -1,13 +1,0 @@
-type Props = {};
-
-const NotFoundView: React.FC<Props> = () => {
-  return (
-    <div>
-      <div className="flex justify-center">
-        <h1>Page not found!</h1>
-      </div>
-    </div>
-  );
-};
-
-export default NotFoundView;
