--- src/pages/Tokens/Tokens_Logic.tsx
@@ -1,22 +1,0 @@
-import TokensView from "./Tokens_View";
-
-const Tokens: React.FC = () => {
-  // const _runners: Runner[] = [];
-  // const { track, number } = useParams();
-
-  // const load = async () => {
-  //   const runners: SignedRunnersResponse = await api.getRunners(
-  //     track || "",
-  //     Number(number) || 0
-  //   );
-  //   setResponse(runners);
-  // };
-
-  // useEffect(() => {
-  //   load();
-  // });
-
-  return <TokensView></TokensView>;
-};
-
-export default Tokens;
