+++ src/pages/Vaults.tsx
@@ -8,7 +8,7 @@ import { useWalletModal } from "../providers/WalletModal";
 import { VaultModalState, VaultTransactionType } from "../types/vaults";
 import { PageLayout } from "../components";
 
-const Vaults: React.FC = () => {
+export const Vaults: React.FC = () => {
   const [modal, setModal] = useState<VaultModalState>();
   const vaultHistory = useSubgraphVaults();
   const config = useConfig();
@@ -97,5 +97,3 @@ const Vaults: React.FC = () => {
     </PageLayout>
   );
 };
-
-export default Vaults;
