+++ src/pages/index.ts
@@ -1,0 +1,7 @@
+export * from "./Dashboard";
+export * from "./Bets";
+export * from "./Markets";
+export * from "./Races";
+export * from "./Tokens";
+export * from "./Vaults";
+export * from "./Faucet";
